requirements.txt
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
azure-cosmos==4.3.0
azure-functions==1.23.0
azure-identity==1.22.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.2.2
protobuf==3.20.3
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
pdfminer.six==20221105



